implementation/pam-system/backend/requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
pyjwt==2.8.0
pyotp==2.9.0
requests==2.31.0
apscheduler==3.10.4
python-multipart==0.0.9
email-validator==2.0.0.post2
